Created structure of App and basic fetching logic
The user types in an IP address in the search component which then emits
the value to the Home view which contains the logic for fetching the data
from the IP information API. This data is then passed as a prop to the
InfoCard component where it's rendered. InfoCard still requires styling.
In the future the Home view will also contain the logic for fetching data
from Leaflet's API.